requirements.txt
black==19.10b0
Cerberus==1.3.2
click==7.1.1
colorama==0.4.3
executor==22.0
Jinja2==2.11.2
mypy-extensions==0.4.3
mypy==0.770
pylint==2.4.4
PyYAML==5.3.1
validators==0.14.3
